refactor: add separate component for Snackbar
